Move interval classes from trino-client to trino-spi
