+++ core/trino-main/src/main/java/io/trino/util/DateTimeUtils.java
@@ -13,10 +13,10 @@
  */
 package io.trino.util;
 
-import io.trino.client.IntervalDayTime;
-import io.trino.client.IntervalYearMonth;
 import io.trino.spi.TrinoException;
 import io.trino.spi.type.TimeZoneKey;
+import io.trino.spi.type.interval.IntervalDayTime;
+import io.trino.spi.type.interval.IntervalYearMonth;
 import org.assertj.core.util.VisibleForTesting;
 import org.joda.time.DateTime;
 import org.joda.time.DurationFieldType;
