+++ core/trino-spi/src/main/java/io/trino/spi/type/interval/IntervalDayTime.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.client;
+package io.trino.spi.type.interval;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
